requirements.txt
flask
flask-sqlalchemy
flask-migrate
pymysql
flask-cors
python-dotenv
cryptography